pyproject.toml
[tool.poetry]
name = "qa-guru-5-21"
version = "0.1.0"
description = ""
authors = ["Artyom Samoshin <artyomzhoy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selene = "2.0.0rc4"
allure-pytest = "^2.13.2"
pytest = "^7.4.2"
jsonschema = "^4.19.0"
curlify = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
